+++ frontend/src/pages/Lobby.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useEffect } from 'react';
 import styled from 'styled-components';
 import { ScaledDiv, ScaledSection } from '@styles/styled';
 import GameModeList from '@components/GameModeList';
@@ -17,95 +17,16 @@ import { getParam } from '@utils/common';
 import { JoinLobbyReEmitRequest, JoinLobbyRequest } from '@backend/core/user.dto';
 import { StartRoundEmitRequest } from '@backend/core/game.dto';
 import { onStartGame } from '@game/NetworkServiceUtils';
-import { userMicCamState, userMicCamType } from '@atoms/user';
+import useWebRTC from '@hooks/useWebRTC';
 
 function Lobby() {
     const [userList, setUserList] = useRecoilState(userListState);
-    const [userMicCam, setUserMicCam] = useRecoilState<userMicCamType>(userMicCamState);
     const [setPage] = useMovePage();
     const lobbyId = getParam('id');
     const setRoundInfo = useSetRecoilState<StartRoundEmitRequest>(roundInfoState);
-
-    const selfVideoRef = useRef<HTMLVideoElement>(null);
-    const selfStreamRef = useRef<MediaStream | undefined>();
-    const remoteVideoRef = useRef<HTMLVideoElement>(null);
-    const pcRef = useRef<RTCPeerConnection>();
-
-    const setMicState = (curState: boolean) => {
-        setUserMicCam({ ...userMicCam, isMicOn: curState });
-        if (!selfStreamRef.current) return;
-        selfStreamRef.current.getAudioTracks().forEach((track) => (track.enabled = !track.enabled));
-    };
-
-    const setCamState = (curState: boolean) => {
-        setUserMicCam({ ...userMicCam, isCamOn: curState });
-        if (!selfStreamRef.current) return;
-        selfStreamRef.current.getVideoTracks().forEach((track) => (track.enabled = !track.enabled));
-    };
-
-    async function getMedia() {
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({
-                video: true,
-                audio: false,
-            });
-            selfStreamRef.current = stream;
-
-            if (!selfVideoRef.current) return;
-            selfVideoRef.current.srcObject = stream;
-
-            selfStreamRef.current.getTracks().forEach((track) => {
-                if (!pcRef.current || !selfStreamRef.current) return;
-                pcRef.current.addTrack(track, selfStreamRef.current);
-            });
-
-            if (!pcRef.current) return;
-
-            pcRef.current.onicecandidate = (e) => {
-                if (e.candidate) {
-                    console.log('recv candidate');
-                    NetworkService.emit('ice', { ice: e.candidate, lobbyId });
-                }
-            };
-
-            pcRef.current.ontrack = (e) => {
-                if (remoteVideoRef.current) {
-                    remoteVideoRef.current.srcObject = e.streams[0];
-                }
-            };
-        } catch (err) {
-            console.log(err);
-        }
-    }
-
-    const createOffer = async () => {
-        if (!pcRef.current) return;
-        try {
-            const offer = await pcRef.current.createOffer();
-            void pcRef.current.setLocalDescription(offer);
-            NetworkService.emit('offer', { offer, lobbyId });
-        } catch (e) {
-            console.error(e);
-        }
-    };
-
-    const createAnswer = async (sdp: RTCSessionDescription) => {
-        console.log('createAnswer');
-        if (!pcRef.current) return;
-
-        try {
-            void pcRef.current.setRemoteDescription(sdp);
-            const answerSdp = await pcRef.current.createAnswer();
-            void pcRef.current.setLocalDescription(answerSdp);
-            NetworkService.emit('answer', { answerSdp, lobbyId });
-        } catch (e) {
-            console.error(e);
-        }
-    };
+    const { selfVideoRef, setMicState, setCamState, getMedia, createOffer } = useWebRTC();
 
     useEffect(() => {
-        pcRef.current = new RTCPeerConnection();
-
         const payload: JoinLobbyRequest = { lobbyId };
         NetworkService.emit(
             'join-lobby',
@@ -123,34 +44,10 @@ function Lobby() {
             setUserList(users.map((user) => user.userName));
         });
 
-        NetworkService.on('offer', (sdp: RTCSessionDescription) => {
-            if (!pcRef.current) return;
-            console.log('on offer');
-            void createAnswer(sdp);
-        });
-
-        NetworkService.on('answer', (sdp: RTCSessionDescription) => {
-            if (!pcRef.current) return;
-            console.log('on answer');
-            void pcRef.current.setRemoteDescription(sdp);
-        });
-
-        // eslint-disable-next-line @typescript-eslint/no-misused-promises
-        NetworkService.on('ice', async (ice: RTCIceCandidate) => {
-            if (!pcRef.current) return;
-            console.log('on ice');
-            await pcRef.current.addIceCandidate(ice);
-            console.log(pcRef.current);
-        });
-
         void getMedia();
 
         return () => {
             NetworkService.off('leave-lobby');
-            NetworkService.off('offer');
-            NetworkService.off('answer');
-            NetworkService.off('ice');
-            if (pcRef.current) pcRef.current.close();
         };
     }, []);
